Move installation routines into proper library
